# Conflicts:
#	packages/Webkul/Shop/src/Resources/lang/ar/app.php
#	packages/Webkul/Shop/src/Resources/lang/de/app.php
#	packages/Webkul/Shop/src/Resources/lang/en/app.php
#	packages/Webkul/Shop/src/Resources/lang/es/app.php
#	packages/Webkul/Shop/src/Resources/lang/it/app.php
#	packages/Webkul/Shop/src/Resources/lang/ja/app.php
#	packages/Webkul/Shop/src/Resources/lang/nl/app.php
#	packages/Webkul/Shop/src/Resources/lang/pl/app.php
#	packages/Webkul/Shop/src/Resources/lang/pt_BR/app.php
#	packages/Webkul/Velocity/publishable/assets/js/jquery.ez-plus.js
#	packages/Webkul/Velocity/publishable/assets/js/velocity.js
#	packages/Webkul/Velocity/src/Resources/assets/js/app.js

+++ packages/Webkul/Shop/src/Resources/views/guest/compare/compare-products.blade.php
@@ -210,6 +210,10 @@
                 },
 
                 'removeProductCompare': function (productId) {
+                    if (productId == 'all' && ! confirm('{{ __('shop::app.customer.compare.confirm-remove-all') }}')) {
+                        return;
+                    }
+
                     if (this.isCustomer) {
                         this.$http.delete(`${this.baseUrl}/comparison?productId=${productId}`)
                         .then(response => {
